Sample Incus init preseed file (btrfs version)
This is a demo file. When pulled on EC2 or any VM/Bare Metal, replace ipv4.address/ipv6.address, NAT, storage_pools size, etc. Using this configuration file may cause pain. This file acts as a placeholder for ansilbe. A better solution to is to use jinja2 but I'm lazy right now. It should work with both LXD and Incus